+++ src/pipelines/features/PipelineTemplates/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PipelineTemplates";
