--- src/pipelines/features/PipelineTemplatesTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./PipelineTemplateTable";
